whatsapp_bot/requirements.txt
fastapi
requests
python-dotenv
uvicorn
pydub
asyncio
google-api-core
google-auth
google-cloud-core
google-cloud-speech
google-cloud-texttospeech
google-cloud-translate
googleapis-common-protos
pymongo
grpcio
pandas
llama-index
llama-index-vector-stores-lancedb
typing
azure-ai-textanalytics
azure-cognitiveservices-speech==1.28.0
azure-common
azure-core
openai
pydantic-settings
sentry-sdk[fastapi]